DOC: Autoreformat all docstrings.
Surprisingly not many changes, most of them are adding leading */** to
args, kwargs which is not clear according to numpydoc but seem like the
thigs to do, plus it's already done in a few other places.

A couple of spaces before `:`, that semantically matter. Numpydoc
improperly parse the types if there is no space before the :,
a documented parameter that does not exist on the function, and
some casing of section names.

There are a couple errors that this was not able to autofix (like
missing blank line before `Returns` section leading Numpydoc to
inpterprete "Returns" as a parameter of the function in
   xarray/core/dataset.py:roll  extra parameters: {'-------', 'rolled', 'Returns'}