Add mode="r+" for to_zarr and improve to_zarr performance
`mode="r+"` only allows for modifying pre-existing array values in a
Zarr store. This makes it a safer default `mode` when doing a limited
`region` write. It also offers a nice performance bonus when using
consolidated metadata, because the store to modify can be opened in
"consolidated" mode.

This PR includes several related changes to `to_zarr()`:

1. It adds support for the new `mode="r+"`.
2. `consolidated=True` in `to_zarr()` now means "open in consolidated
   mode" if using using `mode="r+"`, instead of "write in consolidated
   mode" (which would not make sense for r+).
3. It allows setting `consolidated=True` when using `region`, mostly
   for the sake of fast store opening with r+.
4. Validation in `to_zarr()` has been reorganized to always use the
   _existing_ Zarr group, rather than re-opening zar stores from
   scratch, which could require additional network requests.
5. Incidentally, I've renamed the `ZarrStore.ds` attribute to
   `ZarrStore.zarr_group`, which is a much more descriptive name.

These changes gave me a ~5x boost in write performance in a large
parallel job making use of `to_zarr` with `region`.