Attempt to improve canonical test ds
As discussed in recent issues, I'm trying to create a canonical test ds
that we can base most tests off. We used to do this more often, but it's
less consistent now; and arguably the existing canonical ds could be
improved.

The advantages of the "canonical ds" effort are that we can have a
dataset for which operations should work — e.g. with encoding / attrs /
dtypes / MultiIndexes / etc — without having to remember these for each
operation, and there's a path to being able to abstract over backing
stores, like we've started in `test_dataarray.py`.

But this became somewhat of a yak shave — now that dims are ordered,
this affects test results. So I added the ability to change dimension
order in transpose, but the existing dimension order was very specific
by data variable, so still lots of tests break.

I'm pausing on this, but open to ideas. To handle the changing test
results, I could move the "repr"-based tests to something like
pytest-regtest (which I've used myself for a while and it's pretty good,
though uses file snapshots rather than inline), or doctests
(pytest-accept!); otherwise changing them is very manual.

Also open to feedback that this isn't somewhere to focus on
consolidation, and letting tests evolve as they have been is more than
fine.