Internal refactor of label-based data selection (#5322)
* xindexes also returns multi-index levels as keys

* wip: move label selection into PandasIndex

add Index.query() method
split pd.Index vs. pd.MultiIndex logic

* Revert "xindexes also returns multi-index levels as keys"

This reverts commit 261fb7880523db79b5bc3715477374d089e9a574.

Let's keep this for later. There are too many places in Xarray that
assume that xindexes keys are dimension names.

* fix broken tests

* remove old code + move/update tests

* remove duplicate function

* add PandasMultiIndex class + refactor query impl

* remove PandasIndex.from_variables for now

Add it later in the refactoring when it will be needed elsewhere (e.g., in ``set_index``).

* fix broken tests

Is this what we want?

* prevent loading values for xarray objs in slice

* update what's new