refactor concat
Summary:

- Add the ``Index.concat()`` class method with implementations for
  ``PandasIndex`` and ``PandasMultiIndex`` (adapted from
  ``IndexVariable.concat()``).

- Use ``Index.concat()`` to create new indexes (and their corresponding
  variables) in ``_dataset_concat``. Fallback to variable concatenation
  (with default index) when no implementation is given for
  ``Index.concat`` or when no all the variables have an index.

- Refactored merging the other variables (also merge the indexes) in
  ``_dataset_concat``.

This refactor is incomplete. It should work with Pandas (multi-)indexes
but it is likely that it won't work with meta-indexes involving multiple
dimensions. We probably need to update ``_calc_concat_over`` to take
into account such meta-indexes and their related coordinates.

Other limitation (?) we use here the index of the 1st dataset for the
concatenation (i.e., ``Index.concat``). No specific check is made on
the type and/or coordinates of the other indexes.