to_stacked_array: do not coerce level dtypes
If the intent was to propagate the original coordinate dtypes, this is
now handled by PandasMultiIndex.stack.

This fixes the case where multi-index ``.levels`` do not return labels
with "nan" values (if any) but where the coordinate does, which resulted
in dtype mismatch between the coordinate variable dtype (e.g., coerced
to int64) and the actual label values (float64 due to nan values),
eventually causing numpy errors trying to coerce nan value to int64 when
indexing the coordinate.