Test with warning when loading Zarr with empty dimension with chunks
If an array has zero size (due to an empty dimension), it is saved as a
single chunk regardless of Dask chunking on other dimensions (#5742).
If the `chunks` parameter is provided for other dimensions when loading
the Zarr file, xarray gives a warning about potentially degraded
performance from splitting the single chunk.

When the array has zero size, this warning seems inappropriate because:

- performance degradation on an empty array should be negligible.
- we don't always know if one of the dimensions is empty until loading.
  I would use the `chunks` parameter for dimensions that have known
  chunksize (to specify some multiple of that chunksize), but this only
  works without warning when the array is nonempty.