Add link to xarray binder to readme
I'm not sure this is the best place to put a link, but binder is great,
and we could advertise it more prominently (I couldn't find a hit to
`binder` in the docs either — probably it's under something else?). Any
thoughts on how to do that?

Just now, my M1 Mac seems to have its netcdf installation failing, so I used this
to repro an example — very fast & easy!