Add get_array to lazy indexing array types.
This returns the underlying array type instead of always casting
to np.array. This is necessary for Zarr stores where the
Zarr Array wraps a cupy array (for example kvikio.zarr.GDSStoree).
In that case, we cannot call np.asarray because __array__ is
expected to always return a numpy array.

We use get_array in Variable.data to make sure we don't load arrays
from such GDSStores.