Remove code used to support h5py<2.10.0
It seems that the relevant issue was fixed in 2.10.0
https://github.com/h5py/h5py/commit/466181b178c1b8a5bfa6fb8f217319e021f647e0

I'm not sure how far back you want to fix things. I'm hoping to test
this on the CI.

I found this since I've been auditing slowdowns in our codebase, which
has caused me to review much of the reading pipeline.