Avoid one call to len when getting ndim
I admit this is a super micro optimization but it avoids in certain
cases the creation of a tuple, and a call to len on it.

I hit this as I was trying to understand why Variable indexing was so
much slower than numpy indexing. It seems that bounds checking in python
is just slower than in C.