Add modules to `check-untyped`
In reviewing https://github.com/pydata/xarray/pull/8241, I realize that we actually want `check-untyped-defs`, which is a bit less strict, but lets us add some more modules on. I did have to add a couple of ignores, think it's a reasonable tradeoff to add big modules like `computation` on.

Errors with this enabled are actual type errors, not just `mypy` pedanticness, so would be good to get as much as possible into this list...