ADD: objects can override is_dict_like()
By setting obj._xarray_dict_like_, any object obj can now tell xarray to treat it like it is (or is not) a dict. This is useful if using custom objects for coordinates, when those objects store information in a dict-like format, but that information is not related to xarray and there's no reason for xarray to treat the objects as dicts.