Keep attributes for "bounds" variables
Issue #2921 is about mismatching time units between a time variable and
its "bounds" companion.
However, #2965 does more than fixing #2921, it removes all double
attributes from "bounds" variables which has the undesired side effect
that there is currently no way to save them to netcdf with xarray.
Since the mentioned link is a recommendation and not a hard requirement
for CF compliance, these attributes should be left to the caller to
prepare the dataset variables appropriately if required.
Reduces the amount of surprise that attributes are not written to disk
and fixes #8368.