Updates the example in the doc-string for the Dataset class to be clearer.
The example in the doc-string of the `Dataset` class prior to this commit
uses an example array whose size is `2 x 2 x 3` with the first two dimensions
labeled `"x"` and `"y"` and the final dimension labeled `"time"`. This was
confusing due to the fact that `"x"` and `"y"` are just arbitrary names for
these axes and that no reason is given for the data to be organized in a `2x2x3`
array instead of a `2x2` matrix. This commit clarifies the example.

See issue #8970 for more information.