Docstring and documentation improvement for the Dataset class (#8973)
* Updates the example in the doc-string for the Dataset class to be clearer.

The example in the doc-string of the `Dataset` class prior to this commit
uses an example array whose size is `2 x 2 x 3` with the first two dimensions
labeled `"x"` and `"y"` and the final dimension labeled `"time"`. This was
confusing due to the fact that `"x"` and `"y"` are just arbitrary names for
these axes and that no reason is given for the data to be organized in a `2x2x3`
array instead of a `2x2` matrix. This commit clarifies the example.

See issue #8970 for more information.

* Updates the documentation of the Dataset class to have clearer examples.

These changes to the documentation bring it into alignment with the
changes to the `Dataset` doc-string committed previously.

See issue #8970 for more information.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Adds dataset size reports to the output of the example in the Dataset docstring.

* Fixes the documentation errors in the previous commits.

* Fixes indentation errors in the docs for previous commits.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Maximilian Roos <5635139+max-sixty@users.noreply.github.com>