Disallow passing a DataArray as data into the DataTree constructor
I don't think there's a good reason to support syntax like
`DataTree(DataArray(...))`, when it's easy enough to explicitly
convert with `to_dataset()`. `Dataset(DataArray(...))` currently
raises an error, so this feels a bit inconsistent.

If this was intended for the sake of usability, then I think supporting
a dict (that gets coerced into a Dataset) in the DataTree constructor
would be much helpful. We can consider adding this later.