Remove duplicate coordinates with indexes on sub-trees
This is a _partial_ solution to the duplicate coordinates issue from #9475.

Here we remove all duplicate coordinates between parent and child nodes
with an index (these are already checked for equality via the alignment
check).

Other repeated coordinates (which we cannot automatically check for
equality) are still allowed for now. We will need an alternative
solution for these, as discussed in #9475, but it is less obvious
what the right solution is so I'm holding off on it for now.