Support alternative names for the root node in DataTree.from_dict
`DataTree.from_dict` now supports indicating the root node with "", ".",
"/" or "./".

This makes the handling of the root node a bit more consistent with
handling of other nested paths. For example, we can use relative paths,
which allows for removing special cases for the root from some internal
uses of from_dict.