fix warning from scipy backend guess_can_open on directory
When passing a directory to open_dataset(), the scipy backend fails
and a "RuntimeWarning: 'scipy' fails while guessing" is produced.

This affects me since I'm implementing a backend to read data written
by the adios2 package, whose data "files" are actually a directory.

This tiny patch treats this case just like file not found, that is, the
scipy backend will now return that it cannot open such a "file", but
without raising an exception.