Add `Coordinates.from_xindex` method (+ refactor API doc) (#10000)
* add Coordinates.from_xindex method

* doc: refactor Coordinates API reference

Make it more consistent with ``Dataset``, ``DataArray`` and
``DataTree``. The ``Coordinates`` class is 2nd order compared to the
former ones, but it is public API and useful (for creating coordinates
from indexes and merging coordinates together) so it deserves its own
(expanded) section + summary tables in the API reference doc.

* add tests

* update what's new

* fix doc build?

* docstring tweaks

* doc (api): add missing Coordinates.sizes property

* update what's new (documentation)

* improve docstrings on Coordinates creation

* doc: move what's new entries after last release

---------

Co-authored-by: Deepak Cherian <dcherian@users.noreply.github.com>