[WIP]: Force Zarr coordinate reads to be on the host
zarr-python 3.x supports reading data to host (CPU) memory or device
(GPU) memory. Because coordinates are small and really do need to be on
the host (IIUC because putting them in an Index) then there's no benefit
to reading them to device.

zarr-python includes a global config for whether to use host or device
memory for reads, with `zarr.config.enable_gpu()`. But you can override
that on a per-read basis by passing `prototype` to the getitem call.

This does that for arrays that are coordinates.