Extend `idxmin` and `idxmax` to accept a multiple dimensions
fhis was almost totally done by Claude Code (though our files are super-inefficient, such that it used 17M tokens! which still only costs $8.44 but I guess it could be a couple of orders of magnitude less than that)

my main contribution was trying to find a better way to express the:
```
if (dim is ... or (isinstance(dim, Iterable) and not isinstance(dim, str))) and not isinstance(
    dim, tuple
):
```

we seem to have a lot of functions which touch on this: `infix_dims`, `parse_ordered_dims` (which only seems to be called in our tests?!), `parse_dims_as_tuple`, `parse_dims_as_set`. Let me know if there's a recommended way