Don't skip tests when on a `mypy` branch
For some reason, my local was skipping tests when on a branch named `mypy`, or at least something was setting a `mypy` keyword on the test items.

This changes the check check whether the test has the `mypy` marker set, which is more precise than just looking at the keywords.