Fix type issues from pandas stubs (#10128)
* Fix type issues from pandas stubs

for transparency, claude code also helped here, but I did more auditing & fixing than the idxmax PR (sorry to externalize that to reviewers...)

* .

* Don't skip tests when on a `mypy` branch

For some reason, my local was skipping tests when on a branch named `mypy`, or at least something was setting a `mypy` keyword on the test items.

This changes the check check whether the test has the `mypy` marker set, which is more precise than just looking at the keywords.