Use add_note() to annotate exceptions when encoding fails
We now always add variable names and contents when encoding fails, in
contrast to the current practice where variable names and values are
only sometimes specified, based on the `name` passed into
`VariableCoder.encode()`. This provides better debugging experience for
users.

In the future, we might remove `name` from `VariableCoder.encode()`
because this makes it redundant.

For example, attempting to save a int64 fail in netCDF3 file now shows
something like the following:

    ValueError: could not safely cast array from int64 to int32...
    Raised while encoding variable 'invalid' with value <xarray.Variable (invalid: 1)> Size: 8B
    array([9223372036854775807])

Note that `Exception.add_note()` is a Python 3.11+ feature, so this PR
will need to wait until #10438 is submitted.