Avoid refining parent dimensions in NetCDF files
This changes the way DataTree objects are written to disk in a single
call to `DataTree.to_netcdf()`, as well as netCDF groups written by
passing an explicit `group` to `Dataset.to_netcdf()`.

Conceivably we could only adjust the behavior for `DataTree`, but doing
so for `Dataset` was well felt more consistent to me (and was also
easier to implement).

Fixes: GH10241