Fix unreleased bug in DataTree.to_zarr() with mode='r+'
Writing to zarr with `mode='r+'` was broken by #10625, due to a Zarr
bug (https://github.com/zarr-developers/zarr-python/pull/3428). This
add a work-around that works on older versions of Zarr.

The PR that introduced this bug has not yet appeared in an xarray
release, so there's no need for a release note.