Add assert to CombinedLock
If you have a non-blocking acquire, you have to release all the locks
that you did acquire, but you must not release any of the locks that
have been locked by another thread.

As it is not stored which locks where acquired, the release is not
possible. It is thus not save to use this function without locking.