Add an arithmetic_compat option to xr.set_options, which determines how non-index coordinates of the same name are compared for potential conflicts when performing binary operations.
Previously the behaviour has been fixed at compat='minimal', this allows any of the compat options to be set.

For now the default is still 'minimal', but this sets up arithmetic_compat to be migrated to 'override' alongside the other defaults for compat, as part of use_new_combine_kwarg_defaults. It warns about behaviour which will change under this migration.

A couple of tests which were relying on or testing the arithmetic_compat='minimal' behaviour have been updated to avoid the warning, either by setting arithmetic_compat explicitly or by avoiding a coordinate clash.