Backend fastpath (#10937)
Implements [this proposal](https://github.com/pydata/xarray/issues/10178#issuecomment-2755194660), bypassing the `list_engines` call in the case that an explicit string engine is passed that exists in the standard list. This can shave up to a few seconds off the first `open_dataset` call depending on your env size.