Remove special mapping of `auto` to {} in `open_zarr` (#11010)
Make the handling of `chunks="auto"` consistent between `open_zarr` and `open_dataset(..., engine="zarr")`.

The handling of chunks still differs in `open_zarr` vs `open_dataset(..., engine="zarr")` in that the default in `open_zarr` is to use `chunks={}` and a chunk manager (aka dask) when available in your env. And in `open_dataset` the default is to use `chunks=None` (aka no chunks).

Co-authored-by: Justus Magin <keewis@users.noreply.github.com>

---------

Co-authored-by: Justus Magin <keewis@users.noreply.github.com>